renamed conditions.js to callbacks.js
